wine: Avoid rebuilding on any nixpkgs src change
Previously, the input here was `inputs.nixpkgs`. That meant that *any*
change at all in nixpkgs, to any file, would result in a rebuild. Since
we only depend on files in one dir, use `builtins.path` to filter
nixpkgs to just that dir so that we only depend on changes to that dir,
greatly minimizing unnecessary rebuilds.

Strictly speaking, this is only needed when the path is *not* used in an
import (`callPackage` also being an import). However, there’s little
downside to using this path instead for those as well, and it hopefully
reduces the probability that someone will copy-paste a path using
`inputs.nixpkgs` and use it in a non-import thereby causing the same
problem again.